fix: same MainHeader in different folders
